rename codegen_rust.py to codegen.py
