+++ components/script_bindings/codegen/run.py
@@ -28,7 +28,7 @@ def main():
 
     import WebIDL
     from configuration import Configuration
-    from codegen_rust import CGBindingRoot, CGConcreteBindingRoot
+    from codegen import CGBindingRoot, CGConcreteBindingRoot
 
     parser = WebIDL.Parser(make_dir(os.path.join(out_dir, "cache")))
     webidls = [name for name in os.listdir(webidls_dir) if name.endswith(".webidl")]
@@ -91,7 +91,7 @@ def make_dir(path):
 
 
 def generate(config, name, filename):
-    from codegen_rust import GlobalGenRoots
+    from codegen import GlobalGenRoots
     root = getattr(GlobalGenRoots, name)(config)
     code = root.define()
     with open(filename, "wb") as f:
